refactor: pairRoomMember 패키지 제거 후 pairRoom으로 이동
