+++ backend/src/main/java/site/coduo/pairroom/controller/docs/PairRoomDocs.java
@@ -1,10 +1,13 @@
 package site.coduo.pairroom.controller.docs;
 
+import java.util.List;
+
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -14,6 +17,7 @@ import site.coduo.pairroom.service.dto.PairRoomCreateResponse;
 import site.coduo.pairroom.service.dto.PairRoomReadRequest;
 import site.coduo.pairroom.service.dto.PairRoomReadResponse;
 import site.coduo.pairroom.service.dto.PairRoomStatusUpdateRequest;
+import site.coduo.pairroom.service.dto.PairRoomMemberResponse;
 
 @Tag(name = "페어룸 API")
 public interface PairRoomDocs {
@@ -51,4 +55,17 @@ public interface PairRoomDocs {
             @Parameter(description = "변경할 페어룸 상태", required = true)
             PairRoomStatusUpdateRequest request
     );
+
+    @Operation(summary = "자신의 페어룸을 조회한다.")
+    @ApiResponse(responseCode = "200", description = "페어룸 조회 성공", content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+            schema = @Schema(implementation = PairRoomMemberResponse.class)))
+    ResponseEntity<List<PairRoomMemberResponse>> getPairRooms(
+            @Parameter(
+                    in = ParameterIn.COOKIE,
+                    name = "coduo_whoami",
+                    description = "사용자가 인증에 성공하면 서버에서 발급하는 쿠키",
+                    schema = @Schema(type = "string"),
+                    required = true
+            )
+            String signInToken);
 }
